修改status类型为long

+++ BookInventory/src/main/java/com/lantu/domain/vo/StatusNum.java
@@ -9,8 +9,8 @@ import lombok.Data;
 
 @Data
 public class StatusNum {
-    int matchStatusNum;
-    int notMatchStatusNum;
-    int fixedMatchStatusNum;
-    int errorStatusNum;
+    Long matchStatusNum;
+    Long notMatchStatusNum;
+    Long fixedMatchStatusNum;
+    Long errorStatusNum;
 }
